move App.js to separate folder
