--- src/Contexts.js
@@ -1,6 +1,0 @@
-import { createContext } from 'react';
-
-export const PerformanceContext = createContext({
-  isLowPerformance: false,
-  setIsLowPerformance: () => {},
-});
